hw3: fetch data in App
